server/app/requirements.txt
coreapi==2.1.1
coverage==4.5.3
Django==1.10.5
django-cors-headers==2.0.2
django-debug-toolbar==1.6
django-rest-swagger==2.1.1
djangorestframework==3.5.3
djangorestframework-jwt==1.9.0
itypes==1.1.0
meld3==1.0.2
MySQL-python==1.2.5
openapi-codec==1.2.1
PyJWT==1.4.2
requests==2.13.0
simplejson==3.10.0
sqlparse==0.2.2
supervisor==3.2.0
uritemplate==3.0.0
uWSGI==2.0.14
